fix(lint): clear every clippy warning so the gate can be real
Clippy has never run in CI, so the workspace had accumulated ~35 warnings.
Enabling `-D warnings` without clearing them first would simply turn
develop red, so this clears them.

Mostly machine-applicable rewrites (let-chains, `is_none_or`,
`RangeInclusive::contains`, `div_ceil`, the entry API, `sort_by_key` with
`Reverse`, `io::Error::other`, un-nested `format!` calls). Each was checked
for behaviour equivalence rather than trusted: the `collapsible_match`
rewrites turn `Phase => { if cond {} }` into `Phase if cond => {}`, which
only stays equivalent because the remaining arms cannot match the same
phase and the catch-all is a no-op.

`too_many_arguments` gets `#[allow]` with a reason in all four cases —
two Tauri commands whose argument list *is* the IPC signature, and the two
foul-resolution functions, where the arguments are the resolution context.

Verified against the exact toolchain CI pins (1.95.0), not the local
default: 1.95 carries lints 1.94 does not, and the first pass missed 13 of
them. Both the default build and `--features mcp` are clean, and the
formatting debt on this branch is byte-identical to develop's — no
opportunistic reformatting rode along.

1452 backend tests pass.

+++ src-tauri/src/mcp_server/tools_impl/helpers.rs
@@ -60,7 +60,7 @@ pub(crate) fn age_from_dob(dob: &str, game: &ofm_core::game::Game) -> String {
         Err(_) => return "?".to_string(),
     };
     let ref_date = game.clock.current_date.date_naive();
-    let mut age = i32::from(ref_date.year()) - i32::from(dob_date.year());
+    let mut age = ref_date.year() - dob_date.year();
     if (ref_date.month(), ref_date.day()) < (dob_date.month(), dob_date.day()) {
         age -= 1;
     }
